Implementing a simple material system.
Testing the material approach in the sandbox app.

Improving material system to only hold the pointer of a struct.

Creating light model to test light and materials.

Creating an example of phong light color in the flat shader. OpenGL is done.

Quick fix on shape helper for Geosphere.

Adding the color of the light into the diffuse equation.