Textures and cleanup for DirectX 12
Preparing the code for the render target resource. Changing the sample app to allow directX 12 to render on target.

Render target code almost done. Problem with clear value.

Adding a filter to the debug output to avoid breakpoint for clear render target. Allowing to pass a color to the clearing of the render target.

Adding code to get a texture. Only memory created texture.

Changing to the right format for textures and loading the texture. Creating a smaller texture only for DX12 to test it.

Reflecting the shader to get the texture/sampler information from it.

Finally linking the texture to the drawing flow.