Starting with encrypting scenes. File browser and yaml.
Starting the code for the serialization.

Using YAML with the scene. Establishing architecture, and testing saving and loading a scene.

Modified the ShapeRendererComponentInspector and the ShapeHelper to allow shape selections.

Serializing/Deserializing game objects.

Incrementing the game object serialization effort. Adding TransformComponent serialization.

Creating the shape renderer component serializer.

Fixing the need to call the Material Library in the client side.

Fixing issues with the game objects having the wrong scene after moving the data.

Adding light component serializer.

Loading a scene from the file and showing it on the screen. Fixing issue with light component.