Adding skybox: Opengl
Finishing the test for the skybox with fake cubemap data.

Adding code to load skybox from image and replacing from test texture.

Fixing the issue of not using the projection for the skybox. OpenGL is done.

Allowing the scene settings to change background type and allowing to save/load it.
Fixing minor bug with the camera panning.