Starting to implement shadows in Direct X 12.
Changing the flow of transitions and binding of render targets. Initializing shadow and empty shadow map.

Modifying the Render texture to allow depth writing for shadow map.

Shadows are appearing. But they are not looking correct;

Reviewing code for shadow in DirectX 12. Still not working.

Finally working for Direct X 12 - Shadows