Adding serialization for materials.
Adding deserialization for material.

Allowing to save the editor information in the project.

Testing material serialization.

Loading the scene from the project from the beginning.

Externalizing defaults for materials and loading them in initialization. Grid material not working.

Fixing issue with grid material. Saving the current scene in a variable inside the project object.

Disabling the material property fields in the inspector if the material is default.

Implementing context menu on project structure and creating scenes and materials. Disabling actions for now.

Rewriting file extensions and creating helpers for extensions and folders.

Ability to change the material in the editor. Done.

Finalizing loading and saving materials and scenes in a project.

Pushing defaults.ini template to test in windows.