Adding Metal driver.
Adding the backend for ImGui and extra required classes for the core driver to work.

Adding Metal cpp bridge folder and starting the connection with frameworks and app.

Using GLFW to get a window on metal, instead of doing it manually.

Refactoring the GLFW code for handling events in both windows.

Starting with the basic classes for Metal core.

Integrating the basics for ImGui with apple and metal

Implementing a better way to render by using layer instead of view and delegates.

Renaming file.

Returning the UI manager to its previous state.

Moving the shared parts of the graphics in their place and creating the basics for shaders.

Fixing bugs related to memory and UI manager.

Fixing osx and metal backends, based on https://github.com/ocornut/imgui/commit/70cca1eac07aa3809bdd3717253c3754a5b7cfcc

Adding resize to metal driver.

Passing vertex data from CPU to GPU.

Adding index data for the drawing.

Ability to attach uniforms.

Fixing issues with GLFW window and passing camera data.

This function was causing an error.

Modifying the vertex attributes to send a vertex structure with all the information needed. Testing with a cube.

Starting reflection of the metal shader. All classes have created.

Getting the material information from the shader, binding and setting configuration inside shader.

Moving drawable to surface and command queue to a new class.

Creating the vertex and index buffers, and the buffer group. Replacing test code with buffer group to test the class.

Ability to draw two or more objects at the same time

Ability to draw two or more objects at the same time

Finishing the shader code to work with uniform buffers.